bugfix for (at least) Android 2.3.3
It turned out that Android 2.3.3 (tested on a device and an emulator)
doesn't like calls to dlsym nor to dlopen in a shared library constructor as
we did it here before. The call to dl* did never return, it got stuck
somewhere in the linker code...

Solution (should be, untested till now) to resolve the symbol later when
it's actually used in capset(). While I havn't had time to test this yet
it should work as dlsym and friens are generally supported in Android.