implement missing mips impl + bugfixing
One thing is notable here: Emulators running Android 4.1.2 (API 16) with kernel
2.6.29 show an interesting behavior on the capability side:

We are not able to set our permitted capability set to 0xFFFFFFFF because the
CAP_SETCAP capbilitit is missing (0x100), init, zygote and sh (all running as
root) have the permitted capabilities set to 0xFFFFFEFF. I couldn't figure out
why this is the case, I would guess this is something in the kernel (found some
commits messing with capabilities, but nothing concrete).

Obviuously calls to set permitted to 0xFFFFFFFF will fail, the fix is to set it
everywhere to 0xFFFFFEFF, as it succeeds everywhere and doesn't do harm at all.
We never want to give ourself more capabilities (there is no 'more'), and we
don't want to give capabilities to other people (which is not allowed unless
you have CAP_SETCAP set to 1).